update
git-svn-id: http://svn.tudos.org/repos/oc/tudos/trunk@26 d050ee49-bd90-4346-b210-929a50b99cfc

+++ kernel/fiasco/src/jdb/ia32/jdb_kern_info-ia32-amd64.cpp
@@ -38,7 +38,9 @@ Jdb_kern_info_idt::show()
   Idt::get (&idt_pseudo);
 
   printf("idt base="L4_PTR_FMT"  limit=%04x (%04x bytes)\n",
-	 idt_pseudo.base(), (idt_pseudo.limit()+1)/sizeof(Idt_entry), idt_pseudo.limit()+1);
+         idt_pseudo.base(),
+         (unsigned)((idt_pseudo.limit() + 1) / sizeof(Idt_entry)),
+         idt_pseudo.limit() + 1);
   if (!Jdb_core::new_line(line))
     return;
